Add parsing capability with json files + examples
It won't be viable to manually recompile everytime the particle
generation parameters are changing. Consequently, I added a small
parsing using an open-source JSON library. I refactored the code to use
these parameters everywhere. The syntax of the parameter is still a bit
messy, but this will be refactorable in the future quite easily. This at
least allows the user to drive the simulation from the outside.